refactor(window): move scroll/viewport/recenter handlers to impl Window
`handle_scroll_event`, `handle_set_viewport_event`, and
`handle_recenter_event` are pure per-window concerns — every
self.windows.get(&self.active_window)... access in their bodies was
just a long-form route to per-window splits/buffers/scroll-sync
state.  Moving them to `impl Window` lets us ditch those chains and
gives us direct disjoint sub-field borrows.

Read tab_size from `self.resources.config` instead of Editor's config.
Editor-side callers (event_apply, action_events) chain through
`self.active_window_mut().handle_*` so the external surface is
unchanged.
