+++ crates/fresh-editor/src/app/event_apply.rs
@@ -11,10 +11,10 @@
 //! - `trigger_plugin_hooks_for_event` — broadcast hook notifications
 //!   to plugins after an event applies.
 //!
-//! Plus three "scroll/viewport event" handlers that bypass the buffer
-//! entirely: handle_scroll_event, handle_set_viewport_event,
-//! handle_recenter_event. And a small invalidate_layouts_for_buffer
-//! helper.
+//! The "scroll/viewport event" handlers (handle_scroll_event,
+//! handle_set_viewport_event, handle_recenter_event) and the small
+//! `invalidate_layouts_for_buffer` helper now live on `impl Window`
+//! since they're entirely per-window concerns.
 
 use lsp_types::TextDocumentContentChangeEvent;
 
@@ -48,15 +48,16 @@ impl Editor {
         // This properly separates Buffer state from View state
         match event {
             Event::Scroll { line_offset } => {
-                self.handle_scroll_event(*line_offset);
+                self.active_window_mut().handle_scroll_event(*line_offset);
                 return;
             }
             Event::SetViewport { top_line } => {
-                self.handle_set_viewport_event(*top_line);
+                self.active_window_mut()
+                    .handle_set_viewport_event(*top_line);
                 return;
             }
             Event::Recenter => {
-                self.handle_recenter_event();
+                self.active_window_mut().handle_recenter_event();
                 return;
             }
             _ => {}
@@ -692,367 +693,6 @@ impl Editor {
         // and the plugin's async refreshLines() handles span-level changes.
         if cursor_changed_lines {
             self.handle_refresh_lines(buffer_id);
-        }
-    }
-
-    /// Handle scroll events using the SplitViewState's viewport
-    ///
-    /// View events (like Scroll) go to SplitViewState, not EditorState.
-    /// This correctly handles scroll limits when view transforms inject headers.
-    /// Also syncs to EditorState.viewport for the active split (used in rendering).
-    pub(super) fn handle_scroll_event(&mut self, line_offset: isize) {
-        use crate::view::ui::view_pipeline::ViewLineIterator;
-
-        let active_split = self
-            .windows
-            .get(&self.active_window)
-            .and_then(|w| w.splits.as_ref())
-            .map(|(mgr, _)| mgr)
-            .expect("active window must have a populated split layout")
-            .active_split();
-
-        // Check if this split is in a scroll sync group (anchor-based sync for diffs)
-        // Mark both splits to skip ensure_visible so cursor doesn't override scroll
-        // The sync_scroll_groups() at render time will sync the other split
-        if let Some(group) = self
-            .active_window()
-            .scroll_sync_manager
-            .find_group_for_split(active_split.into())
-        {
-            let left = group.left_split;
-            let right = group.right_split;
-            if let Some(vs) = self
-                .windows
-                .get_mut(&self.active_window)
-                .and_then(|w| w.split_view_states_mut())
-                .expect("active window must have a populated split layout")
-                .get_mut(&LeafId(left))
-            {
-                vs.viewport.set_skip_ensure_visible();
-            }
-            if let Some(vs) = self
-                .windows
-                .get_mut(&self.active_window)
-                .and_then(|w| w.split_view_states_mut())
-                .expect("active window must have a populated split layout")
-                .get_mut(&LeafId(right))
-            {
-                vs.viewport.set_skip_ensure_visible();
-            }
-            // Continue to scroll the active split normally below
-        }
-
-        // Fall back to simple sync_group (same delta to all splits)
-        let sync_group = self
-            .windows
-            .get(&self.active_window)
-            .and_then(|w| w.splits.as_ref())
-            .map(|(_, vs)| vs)
-            .expect("active window must have a populated split layout")
-            .get(&active_split)
-            .and_then(|vs| vs.sync_group);
-        let splits_to_scroll = if let Some(group_id) = sync_group {
-            self.windows
-                .get(&self.active_window)
-                .and_then(|w| w.splits.as_ref())
-                .map(|(mgr, _)| mgr)
-                .expect("active window must have a populated split layout")
-                .get_splits_in_group(
-                    group_id,
-                    self.windows
-                        .get(&self.active_window)
-                        .and_then(|w| w.splits.as_ref())
-                        .map(|(_, vs)| vs)
-                        .expect("active window must have a populated split layout"),
-                )
-        } else {
-            vec![active_split]
-        };
-
-        for split_id in splits_to_scroll {
-            let buffer_id = if let Some(id) = self
-                .windows
-                .get(&self.active_window)
-                .and_then(|w| w.splits.as_ref())
-                .map(|(mgr, _)| mgr)
-                .expect("active window must have a populated split layout")
-                .buffer_for_split(split_id)
-            {
-                id
-            } else {
-                continue;
-            };
-            let tab_size = self.config.editor.tab_size;
-
-            // Get view_transform tokens from SplitViewState (if any)
-            let view_transform_tokens = self
-                .windows
-                .get(&self.active_window)
-                .and_then(|w| w.splits.as_ref())
-                .map(|(_, vs)| vs)
-                .expect("active window must have a populated split layout")
-                .get(&split_id)
-                .and_then(|vs| vs.view_transform.as_ref())
-                .map(|vt| vt.tokens.clone());
-
-            // Get mutable references to both buffer and view state via
-            // a single window borrow split into disjoint sub-fields.
-            let active_id = self.active_window;
-            let window = self
-                .windows
-                .get_mut(&active_id)
-                .expect("active window must exist");
-            if let Some(state) = window.buffers.get_mut(&buffer_id) {
-                let soft_breaks = state.collect_soft_break_positions();
-                let virtual_lines = state.collect_virtual_line_positions();
-                let buffer = &mut state.buffer;
-                if let Some(view_state) = window
-                    .splits
-                    .as_mut()
-                    .expect("active window must have a populated split layout")
-                    .1
-                    .get_mut(&split_id)
-                {
-                    if let Some(tokens) = view_transform_tokens {
-                        // Use view-aware scrolling with the transform's tokens
-                        let view_lines: Vec<_> =
-                            ViewLineIterator::new(&tokens, false, false, tab_size, false).collect();
-                        view_state
-                            .viewport
-                            .scroll_view_lines(&view_lines, line_offset);
-                    } else {
-                        // No view transform - use traditional buffer-based scrolling
-                        if line_offset > 0 {
-                            view_state.viewport.scroll_down(
-                                buffer,
-                                &soft_breaks,
-                                &virtual_lines,
-                                line_offset as usize,
-                            );
-                        } else {
-                            view_state.viewport.scroll_up(
-                                buffer,
-                                &soft_breaks,
-                                &virtual_lines,
-                                line_offset.unsigned_abs(),
-                            );
-                        }
-                    }
-                    // Mark to skip ensure_visible on next render so the scroll isn't undone
-                    view_state.viewport.set_skip_ensure_visible();
-                }
-            }
-        }
-    }
-
-    /// Handle SetViewport event using SplitViewState's viewport
-    fn handle_set_viewport_event(&mut self, top_line: usize) {
-        let active_split = self
-            .windows
-            .get(&self.active_window)
-            .and_then(|w| w.splits.as_ref())
-            .map(|(mgr, _)| mgr)
-            .expect("active window must have a populated split layout")
-            .active_split();
-
-        // Check if this split is in a scroll sync group (anchor-based sync for diffs)
-        // If so, set the group's scroll_line and let render sync the viewports
-        if self
-            .active_window()
-            .scroll_sync_manager
-            .is_split_synced(active_split.into())
-        {
-            if let Some(group) = self
-                .active_window_mut()
-                .scroll_sync_manager
-                .find_group_for_split_mut(active_split.into())
-            {
-                // Convert line to left buffer space if coming from right split
-                let scroll_line = if group.is_left_split(active_split.into()) {
-                    top_line
-                } else {
-                    group.right_to_left_line(top_line)
-                };
-                group.set_scroll_line(scroll_line);
-            }
-
-            // Mark both splits to skip ensure_visible
-            if let Some(group) = self
-                .active_window()
-                .scroll_sync_manager
-                .find_group_for_split(active_split.into())
-            {
-                let left = group.left_split;
-                let right = group.right_split;
-                if let Some(vs) = self
-                    .windows
-                    .get_mut(&self.active_window)
-                    .and_then(|w| w.split_view_states_mut())
-                    .expect("active window must have a populated split layout")
-                    .get_mut(&LeafId(left))
-                {
-                    vs.viewport.set_skip_ensure_visible();
-                }
-                if let Some(vs) = self
-                    .windows
-                    .get_mut(&self.active_window)
-                    .and_then(|w| w.split_view_states_mut())
-                    .expect("active window must have a populated split layout")
-                    .get_mut(&LeafId(right))
-                {
-                    vs.viewport.set_skip_ensure_visible();
-                }
-            }
-            return;
-        }
-
-        // Fall back to simple sync_group (same line to all splits)
-        let sync_group = self
-            .windows
-            .get(&self.active_window)
-            .and_then(|w| w.splits.as_ref())
-            .map(|(_, vs)| vs)
-            .expect("active window must have a populated split layout")
-            .get(&active_split)
-            .and_then(|vs| vs.sync_group);
-        let splits_to_scroll = if let Some(group_id) = sync_group {
-            self.windows
-                .get(&self.active_window)
-                .and_then(|w| w.splits.as_ref())
-                .map(|(mgr, _)| mgr)
-                .expect("active window must have a populated split layout")
-                .get_splits_in_group(
-                    group_id,
-                    self.windows
-                        .get(&self.active_window)
-                        .and_then(|w| w.splits.as_ref())
-                        .map(|(_, vs)| vs)
-                        .expect("active window must have a populated split layout"),
-                )
-        } else {
-            vec![active_split]
-        };
-
-        for split_id in splits_to_scroll {
-            let buffer_id = if let Some(id) = self
-                .windows
-                .get(&self.active_window)
-                .and_then(|w| w.splits.as_ref())
-                .map(|(mgr, _)| mgr)
-                .expect("active window must have a populated split layout")
-                .buffer_for_split(split_id)
-            {
-                id
-            } else {
-                continue;
-            };
-
-            let active_id = self.active_window;
-            let window = self
-                .windows
-                .get_mut(&active_id)
-                .expect("active window must exist");
-            if let Some(state) = window.buffers.get_mut(&buffer_id) {
-                let buffer = &mut state.buffer;
-                if let Some(view_state) = window
-                    .splits
-                    .as_mut()
-                    .expect("active window must have a populated split layout")
-                    .1
-                    .get_mut(&split_id)
-                {
-                    view_state.viewport.scroll_to(buffer, top_line);
-                    // Mark to skip ensure_visible on next render so the scroll isn't undone
-                    view_state.viewport.set_skip_ensure_visible();
-                }
-            }
-        }
-    }
-
-    /// Handle Recenter event using SplitViewState's viewport
-    fn handle_recenter_event(&mut self) {
-        let active_split = self
-            .windows
-            .get(&self.active_window)
-            .and_then(|w| w.splits.as_ref())
-            .map(|(mgr, _)| mgr)
-            .expect("active window must have a populated split layout")
-            .active_split();
-
-        // Find other splits in the same sync group if any
-        let sync_group = self
-            .windows
-            .get(&self.active_window)
-            .and_then(|w| w.splits.as_ref())
-            .map(|(_, vs)| vs)
-            .expect("active window must have a populated split layout")
-            .get(&active_split)
-            .and_then(|vs| vs.sync_group);
-        let splits_to_recenter = if let Some(group_id) = sync_group {
-            self.windows
-                .get(&self.active_window)
-                .and_then(|w| w.splits.as_ref())
-                .map(|(mgr, _)| mgr)
-                .expect("active window must have a populated split layout")
-                .get_splits_in_group(
-                    group_id,
-                    self.windows
-                        .get(&self.active_window)
-                        .and_then(|w| w.splits.as_ref())
-                        .map(|(_, vs)| vs)
-                        .expect("active window must have a populated split layout"),
-                )
-        } else {
-            vec![active_split]
-        };
-
-        for split_id in splits_to_recenter {
-            let buffer_id = if let Some(id) = self
-                .windows
-                .get(&self.active_window)
-                .and_then(|w| w.splits.as_ref())
-                .map(|(mgr, _)| mgr)
-                .expect("active window must have a populated split layout")
-                .buffer_for_split(split_id)
-            {
-                id
-            } else {
-                continue;
-            };
-
-            let active_id = self.active_window;
-            let window = self
-                .windows
-                .get_mut(&active_id)
-                .expect("active window must exist");
-            if let Some(state) = window.buffers.get_mut(&buffer_id) {
-                let buffer = &mut state.buffer;
-                let view_state = window
-                    .splits
-                    .as_mut()
-                    .expect("active window must have a populated split layout")
-                    .1
-                    .get_mut(&split_id);
-
-                if let Some(view_state) = view_state {
-                    // Recenter viewport on cursor
-                    let cursor = *view_state.cursors.primary();
-                    let viewport_height = view_state.viewport.visible_line_count();
-                    let target_rows_from_top = viewport_height / 2;
-
-                    // Move backwards from cursor position target_rows_from_top lines
-                    let mut iter = buffer.line_iterator(cursor.position, 80);
-                    for _ in 0..target_rows_from_top {
-                        if iter.prev().is_none() {
-                            break;
-                        }
-                    }
-                    let new_top_byte = iter.current_position();
-                    view_state.viewport.top_byte = new_top_byte;
-                    // Mark to skip ensure_visible on next render so the scroll isn't undone
-                    view_state.viewport.set_skip_ensure_visible();
-                }
-            }
         }
     }
